% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_crs.R
\name{align_crs}
\alias{align_crs}
\title{Make sure the velox raster and the GPS spdf have the same CRS}
\usage{
align_crs(veloxRasterList, spdf)
}
\arguments{
\item{veloxRasterList}{A list of raster in velox format.}

\item{spdf}{A SpatialPointsDataFrame from the sp package.}
}
\value{
An spdf with the same format as the velox raster data.
}
\description{
Make sure the velox raster and the GPS spdf have the same CRS
}
\examples{
add(1, 1)
add(10, 1)
notes: This should only be run once for the entire operation and thus it should be done before entering the extraction phase.
}
