% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_soil_weights.R
\name{apply_soil_weights}
\alias{apply_soil_weights}
\title{Creates full raster layers for each soil component applying weights to each}
\usage{
apply_soil_weights(raster_list, weight_list = list(5/30, 10/30, 15/30))
}
\arguments{
\item{raster_list}{a list of raster data for each soil property to which to apply the weights}

\item{weight_list}{the list of weights to apply to those soil layers. Defaults specified and shouldn't need to be altered.}
}
\value{
a single raster for each soil component with appropriate weights applied
}
\description{
Creates full raster layers for each soil component applying weights to each
}
\examples{
\dontrun{apply_soil_weights(raster_list, weight_list = list(5/30, 10/30, 15/30))}
}
