% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_soil_units.R
\name{convert_soil_units}
\alias{convert_soil_units}
\title{Creates full raster layers for each soil component, applying weights to each}
\usage{
convert_soil_units(soil_df, variable)
}
\arguments{
\item{soil_df}{a list of raster data for each soil property to which to apply the weights}

\item{variable}{the variable to transform. This function is SUPER SIMPLISTIC.}
}
\value{
the output data set that converts the raw units to more user friendly interpretations
}
\description{
Creates full raster layers for each soil component, applying weights to each
}
\note{
This function will live within the master function call to share the output directly.
}
\examples{
\dontrun{convert_soil_units(soil_data, "ph")}
}
