% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_raster_to_country.R
\name{crop_raster_to_country}
\alias{crop_raster_to_country}
\title{Take the velox raster and crop the velox objects down to the area where the SPDF points are}
\usage{
crop_raster_to_country(raster_list, spdf, country_iso, data_directory)
}
\arguments{
\item{raster_list}{The original list of raster before converting to velox.}

\item{spdf}{SpatialPointsDataFrame to determine extent of interest.}

\item{country_iso}{The country iso code to indicate which country the GPS points are in to know how to trim the data}

\item{data_directory}{The data directory where the polygons are saved.}
}
\value{
raster cropped to the area where soil points are so that we don't care the raw soil rasters through the extract process.
}
\description{
Take the velox raster and crop the velox objects down to the area where the SPDF points are
}
\note{
I've included an if statement in the event there's only one requested soil parameter. There must be a cleaner way to do this.
}
\examples{
\dontrun{crop_raster_to_country("ph_soil_raster.tif", "KEN", "/path/to/directory")}
}
