% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soil_gps.R
\name{extract_soil_gps}
\alias{extract_soil_gps}
\title{Extracts values for GPS df from soil velox raster}
\usage{
extract_soil_gps(velox_raster, spdf, soil_layers)
}
\arguments{
\item{velox_raster}{A list of velox raster from which to extract values}

\item{spdf}{A SpatialPointsDataFrame of the GPS for which we'll be extracting data from the velox raster.}

\item{soil_layers}{the names of the variables being extracted. These are applied to the extracted values.}
}
\value{
A list of velox raster extracts with one point per lon/lat in the spdf
}
\description{
Extracts values for GPS df from soil velox raster
}
\examples{
\dontrun{extract_soil_gps(raster_list, SpatialPointsDataFrame, c("ph", "n"))}
}
