% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country_polygons.R
\name{get_country_polygons}
\alias{get_country_polygons}
\title{checks that country boundaries exist and if they don't downloads them.}
\usage{
get_country_polygons(country_polygon_directory)
}
\arguments{
\item{country_polygon_directory}{the directory where the shape files live}
}
\value{
Clarification that all the countries exist as expected in the designated location
}
\description{
checks that country boundaries exist and if they don't downloads them.
}
\note{
This function is pretty simplistic. It can be simplified by condensing the repeated code. This funciton lets people update the shape files from the GADM website.
}
