% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soil_data.R
\name{get_soil_data}
\alias{get_soil_data}
\title{Executes the full soil extraction process}
\usage{
get_soil_data(spdf, country_iso, soil_layers = c("ph", "carbon",
  "soil_texture", "sand", "silt", "clay", "CEC"), raw_data_directory,
  country_polygon_directory)
}
\arguments{
\item{spdf}{The set of GPS points for which the user wants an export}

\item{country_iso}{The country in which these points are located to trim down the data.}

\item{soil_layers}{The soil variables the user wants returned in the dataframe for those soil points. The soil layers need to be one of the following:
c(ph, carbon, soil_texture, sand, silt, clay, CEC, nitrogen) which are soil ph, carbon, soil texture, sand content, silt content, clay content, CEC, and soil nitrogen}

\item{raw_data_directory}{The location where the data are located to know where to get the raw tif. This defaults to the current working directory.}

\item{country_polygon_directory}{The file directory with the country polygon files. These are available for download with the get_country_polygon() function}
}
\value{
The original spdf data frame along with the extracted soil layers.
}
\description{
Executes the full soil extraction process
}
\note{
This function assumes the data remain named what they are when they're down loaded from the soil grids website.
}
\examples{
import_soil_rasters("af_PHIHOX_T__M", "/Users/mlowes/Google Drive/analyses/soil_grids_raw_data/")
}
