% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_soil_rasters.R
\name{import_soil_rasters}
\alias{import_soil_rasters}
\title{imports appropriate rasters to create soil layers}
\usage{
import_soil_rasters(file_name, dir_name)
}
\arguments{
\item{file_name}{The unique part of the file name to capture the right rasters for that soil}

\item{dir_name}{The directory where the soil raster tif files are stored}
}
\value{
a list of rasters for each soil component
}
\description{
imports appropriate rasters to create soil layers
}
\examples{
\dontrun{import_soil_rasters("af_PHIHOX_T__M",
"/Users/mlowes/Google Drive/analyses/soil_grids_raw_data/")}
}
