% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_map_and_points.R
\name{intersect_map_and_points}
\alias{intersect_map_and_points}
\title{Create intersected map to use to crop velox rasters.}
\usage{
intersect_map_and_points(spdf, country_iso, data_directory)
}
\arguments{
\item{spdf}{SpatialPointsDataFrame to intersect with the polygons}

\item{country_iso}{The country iso code to indicate which country the GPS points are in to know how to trim the data}

\item{data_directory}{the location of the polygon files}
}
\value{
velox raster cropped to the area where soil points are
}
\description{
Create intersected map to use to crop velox rasters.
}
\examples{
\dontrun{apply_soil_weights("ph_soil_layer.tif")}
}
