% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_filter.R
\name{length_filter}
\alias{length_filter}
\title{Breaks list based on subsets with 3 layers and those with one to handle exceptions in the weights and sum portions.}
\usage{
length_filter(raster_list)
}
\arguments{
\item{raster_list}{a list of raster data}
}
\value{
A list with two elements, the first is the set to be weighted or summed, the second is to be combined with the results of the first.
}
\description{
Breaks list based on subsets with 3 layers and those with one to handle exceptions in the weights and sum portions.
}
\examples{
\dontrun{length_filter(list_of_rasters)}
}
