% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_soil_crs.R
\name{match_soil_crs}
\alias{match_soil_crs}
\title{Given SPDF and country where the points are, put the SPDF and country outline
in the same CRS as the soil data which we know is fixed}
\usage{
match_soil_crs(obj_to_convert)
}
\arguments{
\item{obj_to_convert}{The spdf or polygon we need to convert}
}
\value{
that same object but in the correct CRS
}
\description{
Given SPDF and country where the points are, put the SPDF and country outline
in the same CRS as the soil data which we know is fixed
}
\examples{
\dontrun{match_soil_crs(raster_object)}
}
