% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_p_layer.R
\name{prepare_p_layer}
\alias{prepare_p_layer}
\title{Prepares the P data in particular and performs raster repojection to match
the core set.}
\usage{
prepare_p_layer(raw_data_directory, forceUpdate)
}
\arguments{
\item{raw_data_directory}{the location of the file. This should be the in the
same place as the other raw data.}

\item{forceUpdate}{ensures we don't accidentally run this re-projection again. It takes a long time!}
}
\value{
a vector of the file names to get from the data_directory
}
\description{
Prepares the P data in particular and performs raster repojection to match
the core set.
}
\note{
I'll do this once for P and for N so that we don't have to reproject
  the data each time. This will be done in the background once and saved and
  the data will just be available normally through the functions.

  There's no input for the function because it executes a standard data
  transformation.
}
\examples{
\dontrun{layers_to_files(c("carbon", "CEC"))}
}
