% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_and_sum_soil.R
\name{stack_and_sum_soil}
\alias{stack_and_sum_soil}
\title{Takes weighted soil layers, creates stack, and sums values to create single soil raster for that parameter}
\usage{
stack_and_sum_soil(raster_list)
}
\arguments{
\item{raster_list}{a list of weighted soil rasters}
}
\value{
a single raster layer that sums the soil values including the weights
  for the velox extraction stage
}
\description{
Takes weighted soil layers, creates stack, and sums values to create single soil raster for that parameter
}
\note{
This is a pretty simple function but I wanted to keep this separate from the other preparation stages.
}
\examples{
\dontrun{stack_and_sum_soil()}
}
