% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_soil_to_velox.R
\name{convert_soil_to_velox}
\alias{convert_soil_to_velox}
\title{Converts soil raster to velox}
\usage{
convert_soil_to_velox(raster_list)
}
\arguments{
\item{raster_list}{a raster of soil data from the soil grids repository to be converted to raster for extraction}
}
\value{
a velox raster version of the soil data for extraction
}
\description{
Converts soil raster to velox
}
\examples{
\dontrun{convert_soil_to_velox(list("ph_soil_layer.tif", "n_soil_layer.tif"))}
}
